feat: Function Integration
